Cleanup and comment
